Python/requirements.txt
albumentations==1.3.1
csbdeep==0.7.4
Cython==3.0.8
geojson==3.1.0
matplotlib==3.7.4
numpy==1.23.5
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
openpyxl==3.1.2
openslide_python==1.1.2
pandas==2.2.1
Pillow==9.4.0
Pillow==10.2.0
rasterio==1.3.9
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
segmentation_models_pytorch==0.3.3
tifffile==2023.9.26
torch==2.1.2
torchvision==0.16.0
tqdm==4.66.1
valis==0.1
